HDRP Defaults
The most important thing to get from this branch for lighting is the SKY_FOG_PREFAB under the SkyandFog folder in prefabs, put that in every scene. As well as the HDRI used in the default settings. Under Edit - Project settings - HDRP Global settings, ensure that HDRI sky is checked on and that it is using the Sunflowers_puresky_2k texture.